+++ src/test/java/bowling/view/GameStatusTest.java
@@ -1,5 +1,6 @@
-package bowling;
+package bowling.view;
 
+import bowling.GameService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
